is_int_overflow function fixed (return of ft_atol in libft edited from int to long)

+++ lib/Libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/12 19:30:21 by oozsertt          #+#    #+#             */
-/*   Updated: 2021/11/14 18:18:39 by user42           ###   ########.fr       */
+/*   Updated: 2021/11/15 12:44:39 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ typedef int	t_bool;
 
 int		ft_atoi_base(char *str, char *base);
 int		ft_atoi(const char *str);
-int		ft_atol(const char *str);
+long	ft_atol(const char *str);
 char	*ft_convert_base(char *nbr, char *base_from, char *base_to);
 char	*ft_convert_fullbinary(char *nbr, char *decimal, char *binary);
 t_bool	ft_isalnum(int c);
